+++ app/services/event_service.rb
@@ -71,7 +71,16 @@ class EventService
       events.each do |e|
         method = e.destroyable? ? :destroy : :soft_destroy!
         # we use double negation because destroy can return either a boolean (false) or an Event (in case of delete success)
-        results.push status: !!e.send(method), event: e # rubocop:disable Style/DoubleNegation
+
+        ActiveRecord::Base.transaction do
+          status = !!e.send(method)
+          if status
+            e.reservations.preload(:slots_reservations).map(&:slots_reservations).flatten.map do |slots_reservation|
+              SlotsReservationsService.cancel(slots_reservation)
+            end
+          end
+          results.push status: status, event: e # rubocop:disable Style/DoubleNegation
+        end
       end
       results
     end