+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - updates spanish translations
 - Fix a bug: avoids crash due to oidc config with scope = nil
 - Fix a bug: unable to see value for input group with long label on eventModal
+- Improvement: when deleting an event, all reservations are canceled
 
 ## v6.3.0 2023 November 3